requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
openai==1.58.1
httpx==0.28.0
jinja2==3.1.3
python-multipart==0.0.6
pyodbc
chromadb
langgraph
pandas
pypdf
python-docx
pdfplumber
openpyxl